+++ src/atompy/_histogram.py
@@ -5,7 +5,6 @@ from numpy.typing import NDArray
 from dataclasses import dataclass
 from . import _io as apio
 from . import _miscellaneous as _misc
-from ._miscellaneous import get_all_dividers, ImshowData, PcolormeshData
 from typing import Literal
 
 
@@ -67,7 +66,7 @@ class Hist1d:
         if old_n % factor != 0:
             raise ValueError(
                 f"Invalid {factor=}. Possible factors for this "
-                f"histogram are {get_all_dividers(old_n)}."
+                f"histogram are {_misc.get_all_dividers(old_n)}."
             )
 
         new_hist = np.empty(self.histogram.size // factor)
@@ -329,7 +328,7 @@ class Hist2d:
         if old_n % factor != 0:
             raise ValueError(
                 f"Invalid {factor=}. Possible factors for this "
-                f"histogram are {get_all_dividers(old_n)}."
+                f"histogram are {_misc.get_all_dividers(old_n)}."
             )
         new_hist = np.empty((self.H.shape[0] // factor, self.H.shape[1]))
         for i in range(new_hist.shape[0]):
@@ -362,7 +361,7 @@ class Hist2d:
         if old_n % factor != 0:
             raise ValueError(
                 f"Invalid {factor=}. Possible factors for this "
-                f"histogram are {get_all_dividers(old_n)}."
+                f"histogram are {_misc.get_all_dividers(old_n)}."
             )
         new_hist = np.empty((self.H.shape[0], self.H.shape[1] // factor))
         for i in range(new_hist.shape[1]):
@@ -376,7 +375,7 @@ class Hist2d:
     @property
     def for_pcolormesh(
         self
-    ) -> PcolormeshData:
+    ) -> _misc.PcolormeshData:
         """
         Return such that it can be plotted using
         :obj:`matplotlib.pyplot.pcolormesh`
@@ -399,12 +398,12 @@ class Hist2d:
             hist = Hist2d(*np.histogram2d(xsamples, ysamples))
             plt.pcolormesh(*hist.for_pcolormesh)
         """
-        return PcolormeshData(self.xedges, self.yedges, self.H.T)
+        return _misc.PcolormeshData(self.xedges, self.yedges, self.H.T)
 
     @property
     def for_imshow(
         self,
-    ) -> ImshowData:
+    ) -> _misc.ImshowData:
         """
         Return corresponding :class:`.ImshowData` object.
 
@@ -449,7 +448,7 @@ class Hist2d:
                  self.xedges.max(),
                  self.yedges.min(),
                  self.yedges.max()]
-        return ImshowData(H_, np.array(edges))
+        return _misc.ImshowData(H_, np.array(edges))
 
     def within_xrange(
         self,