+++ changelog.md
@@ -1,8 +1,8 @@
 # v4.2.0
 - Added examples to documentation
+- Code refactoring
 
 # v4.1.0
-- Add examples to documentation
 
 # v4.0.1
 - Replaced `np.float_` with `np.float64`